Insane conficts resolved; compile errors fixed on state_machine.rs,
everything changed from u64 to uint; fields made public when previously
private .... blah blah blah.

It "works", at least no worse than on any individual branch. The tests
pass. Rejoice.

+++ src/schooner/events/vote.rs
@@ -2,16 +2,16 @@ use uuid::Uuid;
 
 #[deriving(Decodable, Encodable, Eq, Clone, Show)]
 pub struct VoteReq {
-    pub term: uint,
-    pub candidate_id: uint,
-    pub last_log_index: uint,
-    pub last_log_term: uint,
+    pub term: u64,
+    pub candidate_id: u64,
+    pub last_log_index: u64,
+    pub last_log_term: u64,
     pub uuid: Uuid,
 }
 
 #[deriving(Decodable, Encodable, Eq, Clone, Show)]
 pub struct VoteRes {
-    pub term: uint,
+    pub term: u64,
     pub vote_granted: bool,
     pub uuid: Uuid,
 }
